Refactor - Split sites between twitch/youtube
