+++ src/Sites/twitch.tv/Components/ChatInput.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { EmoteComponent } from 'src/Content/Components/EmoteComponent';
 import { EmoteStore } from 'src/Global/EmoteStore';
-import { MessagePatcher } from 'src/Page/Util/MessagePatcher';
+import { MessagePatcher } from 'src/Sites/twitch.tv/Util/MessagePatcher';
 
 /**
  * Chat Input
